Change "export_segments" behavior to expand bounding box to (24, 24) size.
This is different from previous behavior, which resized small segments to the
minimum image size. This excluded relevant surrounding content, which reduced
the interpretability of the segment images.

This solution is a little hacked together, and should be cleaned up when properly
constructing a dataset of segment images. This is purely for demonstration
purposes to explain to Kwang Moo Yi/Alexandra Albu.